+++ lib/src/common/custom_animated_positioned.dart
@@ -2,11 +2,12 @@ import 'package:flutter/material.dart';
 
 /// CustomAnimatedPositioned is almost identical to AnimatedPositioned
 /// only integrating the isBeingDragged property to enable us to set the node being dragged at front
-class CustomAnimatedPositioned extends AnimatedPositioned {
+class CustomAnimatedPositionedDirectional
+    extends AnimatedPositionedDirectional {
   /// Whether the node is being dragged or not
   final bool isBeingDragged;
 
-  const CustomAnimatedPositioned({
+  const CustomAnimatedPositionedDirectional({
     super.key,
     required super.child,
     required super.duration,
@@ -14,9 +15,9 @@ class CustomAnimatedPositioned extends AnimatedPositioned {
     super.curve,
     super.onEnd,
     super.top,
-    super.right,
+    super.start,
     super.bottom,
-    super.left,
+    super.end,
     super.width,
     super.height,
   });